+++ packages/app/lib/locales/app_en.arb
@@ -3,6 +3,7 @@
     "createSightingError": "Something went wrong: {error}",
     "createSightingScreenTitle": "Add Sighting",
     "createSightingSuccess": "Yay, you've successfully added a new sighting to the database",
+    "deleteSighting": "Delete Sighting",
     "editCardCancelButton": "Cancel",
     "editCardSaveButton": "Save",
     "editSpeciesErrorCantRemove": "Species needs to be defined",
@@ -24,17 +25,12 @@
     "imageDeleteAlertCancel": "No",
     "imageDeleteAlertConfirm": "Yes",
     "imageDeleteAlertTitle": "Delete Image",
-    "sightingDeleteAlertBody": "Are you sure you want to delete this sighting?",
-    "sightingDeleteAlertCancel": "No",
-    "sightingDeleteAlertConfirm": "Yes",
-    "sightingDeleteAlertTitle": "Delete Sighting",
-    "sightingDeleteConfirmation": "Sighting deleted.",
     "imageDeleteConfirmation": "Image deleted.",
     "imageLoadingError": "Could not load image",
     "imageMissingError": "No image given",
+    "localNameCardSaveAction": "Save",
     "localNameCardTitle": "Local Name",
     "localNamesCardTitle": "Local Names",
-    "localNameCardSaveAction": "Save",
     "locationAdd": "Do you want to add a GPS location?",
     "locationAddAction": "Add location to image",
     "locationErrorPermissionDenied": "Location permission was denied",
@@ -62,11 +58,13 @@
     "settingsSystemInfoError": "Could not retrieve system informations",
     "settingsSystemInfoSDK": "Android SDK Version",
     "settingsSystemInformation": "System Information",
+    "sightingDeleteAlertBody": "Are you sure you want to delete this sighting?",
+    "sightingDeleteAlertCancel": "Cancel",
+    "sightingDeleteAlertConfirm": "Delete",
+    "sightingDeleteAlertTitle": "Delete Sighting",
+    "sightingDeleteConfirmation": "Sighting deleted.",
     "sightingScreenTitle": "Sighting",
     "sightingUnspecified": "Unknown species",
-    "speciesScreenTitle": "Species",
-    "usedForCardTitle": "Used For",
-    "usedForTextSave": "Save",
     "speciesCardTitle": "Species",
     "speciesDescription": "Description",
     "speciesScreenTitle": "Species",
@@ -78,5 +76,7 @@
     "taxonomyPhylum": "Phylum",
     "taxonomySpecies": "Species",
     "taxonomySubfamily": "Subfamily",
-    "taxonomyTribe": "Tribe"
+    "taxonomyTribe": "Tribe",
+    "usedForCardTitle": "Used For",
+    "usedForTextSave": "Save"
 }
